Lab3 add preempt
