+++ Lab3/kernel/lib/uart.c
@@ -1,5 +1,12 @@
 #include "uart.h"
 
+char tx_buffer[MAX_BUF_SIZE];
+char rx_buffer[MAX_BUF_SIZE];
+unsigned int tx_bf_r_idx = 0;
+unsigned int tx_bf_w_idx = 0;
+unsigned int rx_bf_r_idx = 0;
+unsigned int rx_bf_w_idx = 0;
+
 void uart_init()
 {
     register unsigned int r;
@@ -144,4 +151,99 @@ char uart_get() {
     r=(char)(*AUX_MU_IO);
     /* convert carriage return to newline */
     return r;
+}
+
+void enable_uart_interrupt() {
+    enable_uart_tx_interrupt();
+    enable_uart_rx_interrupt();
+    *ENABLE_IRQS_1 |= 1 << 29;
+}
+
+void enable_uart_tx_interrupt() {
+    *AUX_MU_IER |= 2;
+}
+
+void disable_uart_tx_interrupt() {
+    *AUX_MU_IER &= ~(2);
+}
+
+void enable_uart_rx_interrupt() {
+    *AUX_MU_IER |= 1;
+}
+
+void disable_uart_rx_interrupt() {
+    *AUX_MU_IER &= ~(1);
+}
+
+void uart_tx_interrupt_handler() { // output
+    // buffer empty
+    if (tx_bf_r_idx == tx_bf_w_idx) {
+        disable_uart_tx_interrupt();
+        return;
+    }
+
+    *AUX_MU_IO = (unsigned int)tx_buffer[tx_bf_r_idx];
+    tx_bf_r_idx++;
+    tx_bf_r_idx %= MAX_BUF_SIZE;
+
+    enable_uart_tx_interrupt();
+}
+
+void uart_rx_interrupt_handler() { // input
+    // buffer full -> wait for empty space
+    if ((rx_bf_w_idx + 1) % MAX_BUF_SIZE == rx_bf_r_idx) {
+        disable_uart_rx_interrupt();
+        return;
+    }
+    
+    disable_interrupt();
+    rx_buffer[rx_bf_w_idx] = (char)(*AUX_MU_IO);
+    rx_bf_w_idx++;
+    rx_bf_w_idx %= MAX_BUF_SIZE;
+    enable_interrupt();
+}
+
+void uart_async_putc(char c) {
+    // buffer full -> wait for empty space
+    while ((tx_bf_w_idx + 1) % MAX_BUF_SIZE == tx_bf_r_idx) {
+        enable_uart_tx_interrupt();
+    }
+
+    disable_interrupt();
+    tx_buffer[tx_bf_w_idx] = c;
+    tx_bf_w_idx++;
+    tx_bf_w_idx %= MAX_BUF_SIZE;
+    enable_interrupt();
+
+    enable_uart_tx_interrupt();
+}
+
+char uart_async_getc() {
+    // buffer empty -> wait
+    while (rx_bf_r_idx == rx_bf_w_idx) {
+        enable_uart_rx_interrupt();
+    }
+
+    char c = rx_buffer[rx_bf_r_idx];
+    rx_bf_r_idx++;
+    rx_bf_r_idx %= MAX_BUF_SIZE;
+
+    return c;
+}
+
+int uart_async_printf(char *fmt, ...)
+{
+    __builtin_va_list args;
+    __builtin_va_start(args, fmt);
+    char buf[MAX_BUF_SIZE];
+    char *s = (char *)buf;
+    int count = vsprintf(s, fmt, args);
+    while (*s)
+    {
+        if (*s == '\n')
+            uart_async_putc('\r');
+        uart_async_putc(*s++);
+    }
+    __builtin_va_end(args);
+    return count;
 }
